Improve user interactivity by adding login functionality
A login functionality was added to the 'loginMenu' method in the 'CourseRegistrationSystem' class. Users are prompted to input their username and password. Depending on their class - either 'Student' or 'Advisor' - the respective menu is loaded upon successful login. This enhancement provides personalized user interaction and error handling for non-existent users.