Improve 'readLecturers' method to parse JSON data
Updated the 'readLecturers' method in the 'JSONReader' class to parse a JSON file for lecturer data. The data is then converted into 'Lecturer' objects and added to the department's lecturer list. This update provides a more efficient way to handle and manage lecturer data while reducing error risk due to file not found exceptions.