Improve draft course setting in JSONReader
Moved the draftCourses list instantiation inside the for loop of the `readRequests` method to prevent propagation of draft courses across different student requests. Also, this commit replaces "courseCodes" with "courses" as the attribute name in JSON accounting to changes in the data model. Additionally, a block of code that was commented out was removed since it appeared to be redundant and didn't serve any functionality. This ensures each student's draft courses are distinct and not influenced by previous iterations, improving the accuracy of request processing.